+++ src/main/java/cortana/command/Command.java
@@ -1,3 +1,10 @@
+package cortana.command;
+
+import cortana.exception.CortanaException;
+import cortana.storage.FileHandler;
+import cortana.task.TaskList;
+import cortana.ui.Ui;
+
 import java.io.IOException;
 
 /**
