+++ src/main/java/cortana/command/DeleteCommand.java
@@ -1,3 +1,10 @@
+package cortana.command;
+
+import cortana.exception.CortanaException;
+import cortana.storage.FileHandler;
+import cortana.task.TaskList;
+import cortana.ui.Ui;
+
 import java.io.IOException;
 
 /**
@@ -7,7 +14,7 @@ public class DeleteCommand implements Command {
     private final int taskNumber;
 
     /**
-     * Constructs a DeleteCommand for deleting a specific task.
+     * Constructs a cortana.command.DeleteCommand for deleting a specific task.
      *
      * @param taskNumber The index of the task to delete (1-based)
      */
@@ -17,7 +24,7 @@ public class DeleteCommand implements Command {
 
     /**
      * Executes the command to delete the specified task.
-     * Updates the task list and persists the task deletion using FileHandler.
+     * Updates the task list and persists the task deletion using cortana.storage.FileHandler.
      *
      * @param taskList The list of tasks to modify
      * @param ui The UI for displaying output messages
