+++ src/main/java/cortana/storage/FileHandler.java
@@ -1,3 +1,11 @@
+package cortana.storage;
+
+import cortana.exception.CortanaException;
+import cortana.task.Deadline;
+import cortana.task.Event;
+import cortana.task.TaskList;
+import cortana.task.ToDo;
+
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -21,7 +29,7 @@ public class FileHandler {
     }
 
     /**
-     * Constructs a FileHandler for the given file path.
+     * Constructs a cortana.storage.FileHandler for the given file path.
      *
      * @param filePath the path to the data file
      */
@@ -120,10 +128,10 @@ public class FileHandler {
     }
 
     /**
-     * Loads the tasks stored in the data file into a TaskList object.
+     * Loads the tasks stored in the data file into a cortana.task.TaskList object.
      * Each line in the file is parsed into a corresponding task with their done statuses.
      *
-     * @return the TaskList loaded from the file
+     * @return the cortana.task.TaskList loaded from the file
      * @throws CortanaException if there is an error related to task creation
      * @throws IOException if an I/O error occurs reading the file
      */
@@ -142,10 +150,10 @@ public class FileHandler {
             String markValue = parts[1];
             String taskName = parts[2];
 
-            // for ToDo tasks
+            // for cortana.task.ToDo tasks
             if (taskCommand.equals("T")) {
                 tasks.add(new ToDo(taskName));
-            // for Deadline tasks
+            // for cortana.task.Deadline tasks
             } else if (taskCommand.equals("D")) {
                 String taskBy = parts[3];
                 LocalDateTime byDateTime = LocalDateTime.parse(taskBy, DateTimeFormatter.ofPattern("dd MMM yy HHmm"));
@@ -173,8 +181,8 @@ public class FileHandler {
      * The task string is expected in the format: "[T][ ] Read Book".
      * This is converted and appended as "T | 0 | Read Book".
      *
-     * @param tasks the current TaskList (not used here but may be useful)
-     * @param taskString the string representation of the ToDo task to save
+     * @param tasks the current cortana.task.TaskList (not used here but may be useful)
+     * @param taskString the string representation of the cortana.task.ToDo task to save
      * @throws CortanaException if an I/O error occurs during file writing
      */
     public void saveToDo(TaskList tasks, String taskString) throws CortanaException {
@@ -194,8 +202,8 @@ public class FileHandler {
      * The task string is expected in the format: "[T][ ] Read Book (by: Sunday)".
      * This is converted and appended as "T | 0 | Read Book | Sunday".
      *
-     * @param tasks the current TaskList (not used here but may be useful)
-     * @param taskString the string representation of the Deadline task to save
+     * @param tasks the current cortana.task.TaskList (not used here but may be useful)
+     * @param taskString the string representation of the cortana.task.Deadline task to save
      * @throws CortanaException if an I/O error occurs during file writing
      */
     public void saveDeadline(TaskList tasks, String taskString) throws CortanaException {
@@ -221,8 +229,8 @@ public class FileHandler {
      * The task string is expected in the format: "[T][ ] Read Book (from: Sunday 4pm to: 6pm)".
      * This is converted and appended as "T | 0 | Read Book | Sunday 4pm | 6pm".
      *
-     * @param tasks the current TaskList (not used here but may be useful)
-     * @param taskString the string representation of the Event task to save
+     * @param tasks the current cortana.task.TaskList (not used here but may be useful)
+     * @param taskString the string representation of the cortana.task.Event task to save
      * @throws CortanaException if an I/O error occurs during file writing
      */
     public void saveEvent(TaskList tasks, String taskString) throws CortanaException {
@@ -250,7 +258,7 @@ public class FileHandler {
     /**
      * Updates the done status mark value (0 or 1) on a specific line in the data file.
      *
-     * @param tasks the current TaskList (not used here but may be useful)
+     * @param tasks the current cortana.task.TaskList (not used here but may be useful)
      * @param lineNumber the 1-based line number to update
      * @param markValue the mark value to set ("0" or "1")
      * @throws CortanaException if an I/O error occurs during update
@@ -281,7 +289,7 @@ public class FileHandler {
     /**
      * Deletes a task line from the data file by line number.
      *
-     * @param tasks the current TaskList (not used here but may be useful)
+     * @param tasks the current cortana.task.TaskList (not used here but may be useful)
      * @param lineNumber the 1-based line number of the task to delete
      * @throws CortanaException if an I/O error occurs during delete
      */
